feat: Architecture refactoring
