+++ README.md
@@ -4,10 +4,10 @@ A simple bot for my personal discord server
 
 ## Features
 
-- Update a channel name according to the number of messages inside it
-- Receive YouTube notifications in a specified channels
-    > Created to allow unlimited subscriptions (bots like [Carl](https://carl.gg/) have limited free tier)
- 
-## Credits 
+- [Update a channel name](src/tasks/counter.ts) according to the number of messages inside it
+- [Receive YouTube notifications](src/tasks/youtube.ts) in a specified channels
+  > Created to allow unlimited subscriptions (bots like [Carl](https://carl.gg/) have limited free tier)
+
+## Credits
 
 The YouTube notification feature is highly inspired from [Androz2091/DiscordYoutubeNotifier](https://github.com/Androz2091/DiscordYoutubeNotifier).
