+++ src/tasks/index.ts
@@ -1,0 +1,9 @@
+import setupCounter from "./counter";
+import setupYoutube from "./youtube";
+
+const setupTasks = () => {
+	setupCounter();
+	setupYoutube();
+};
+
+export default setupTasks;
